Don't constrain on scope when looking for roots
When $scoped is set the Node::roots method was breaking since it was
attempting to find nodes with the same scope as $this (which was a
newly instantiated Node with no scope).

This patch introduces a new attribute to the newNestedSetQuery method
to constrain to the current scope, true by default. This is then set
to false when called from the roots method.