Reimplement 'toHierarchy' method.
This method was yielding inconsistent results, even worse with custom sorted
result sets. Also it was *position dependent*, in that it needed the nodes to
nest next to each other (this has been bothering me quite some time).

This new impl is not recursive, not position dependent and respects the
actual ordering the original result set used.